pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "vlui-research"
version = "0.1.0"
description = "Locality-Constrained Spatial Indexing with Provable Cross-Region Interference Bounds"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "matplotlib>=3.7",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "rtree>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-benchmark>=4.0", "black", "mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
